Split FE class into two: Base class hols all functions related to AssembleFEElements -> Element-wise jacobian constructoion, the derived FE class has all other functions related to thge original global assembly routines - Important: NO code was deleted only moved
