+++ feddlib/core/FE/FE_ElementAssembly.hpp
@@ -1,0 +1,8 @@
+#ifndef FE_ElementAssembly_hpp
+#define FE_ElementAssembly_hpp
+#include "FE_ElementAssembly_decl.hpp"
+#ifndef HAVE_EXPLICIT_INSTANTIATION
+    #include "FE_ElementAssembly_def.hpp"
+#endif
+
+#endif // MATRIX_hpp
