+++ feddlib/core/FE/FE_decl.hpp
@@ -1,24 +1,7 @@
 #ifndef FE_DECL_hpp
 #define FE_DECL_hpp
 
-
-#include "feddlib/core/FEDDCore.hpp"
-#include "feddlib/core/General/SmallMatrix.hpp"
-#include "feddlib/core/General/DefaultTypeDefs.hpp"
-#include "feddlib/core/LinearAlgebra/Matrix.hpp"
-#include "feddlib/core/LinearAlgebra/MultiVector.hpp"
-#include "feddlib/core/LinearAlgebra/BlockMatrix.hpp"
-#include "Domain.hpp"
-#include "Helper.hpp"
-#include "sms.hpp"
-#include "feddlib/core/AceFemAssembly/AssembleFE.hpp"
-#include "feddlib/core/AceFemAssembly/specific/AssembleFE_SCI_SMC_Active_Growth_Reorientation.hpp"
-#include "feddlib/core/AceFemAssembly/specific/AssembleFENavierStokes.hpp"
-
-#include "feddlib/core/AceFemAssembly/AssembleFEFactory.hpp"
-
-#include <Teuchos_Array.hpp>
-#include <Teuchos_BLAS.hpp>
+#include "FE_ElementAssembly.hpp"
 
 #include <boost/function.hpp>
 
@@ -44,8 +27,9 @@ class DataElement {
         std::vector<double> hp_;
 };
 
+
 template <class SC = default_sc, class LO = default_lo, class GO = default_go, class NO = default_no>
-class FE {
+class FE: public FE_ElementAssembly<SC,LO,GO,NO> {
   public:
 
     typedef Domain<SC,LO,GO,NO> Domain_Type;
@@ -100,9 +84,6 @@ class FE {
 
 	typedef SmallMatrix<SC> SmallMatrix_Type;
     typedef Teuchos::RCP<SmallMatrix_Type> SmallMatrixPtr_Type;
-
-    DomainConstPtr_vec_Type	domainVec_;
-    Teuchos::RCP<ElementSpec> es_;
 
     /* ###################################################################### */
 
@@ -233,12 +214,6 @@ class FE {
     //                        bool updateHistory=true);
     
     
-    void addFE(DomainConstPtr_Type domain);
-
-    void doSetZeros(double eps = 10*Teuchos::ScalarTraits<SC>::eps());
-
-    void assemblyEmptyMatrix(MatrixPtr_Type &A);
-
     void applyBTinv(vec3D_dbl_ptr_Type& dPhiIn,
                     vec3D_dbl_Type& dPhiOut,
                     SmallMatrix<SC>& Binv);
@@ -282,11 +257,6 @@ class FE {
                                  MatrixPtr_Type &A,
                                  bool callFillComplete = true);
 
-    void assemblyNonlinearLaplace(
-        int dim, std::string FEType, int degree, MultiVectorPtr_Type u_rep,
-        BlockMatrixPtr_Type &A, BlockMultiVectorPtr_Type &resVec,
-        ParameterListPtr_Type params, std::string assembleMode,
-        bool callFillComplete = true, int FELocExternal = -1);
 
     // // Assembling the reaction term of the reaction diffusion equation. Maybe add default function.
 	// void assemblyLinearReactionTerm(int dim,
@@ -542,172 +512,10 @@ class FE {
 
     void epsilonTensor(vec_dbl_Type &basisValues, SmallMatrix<SC> &epsilonValues, int activeDof);
 
-    void assemblyNavierStokes(int dim,
-								std::string FETypeVelocity,
-								std::string FETypePressure,
-								int degree,
-								int dofsVelocity,
-								int dofsPressure,
-								MultiVectorPtr_Type u_rep,
-								MultiVectorPtr_Type p_rep,
-								BlockMatrixPtr_Type &A,
-								BlockMultiVectorPtr_Type &resVec,
-								SmallMatrix_Type coeff,
-								ParameterListPtr_Type params,
-								bool reAssemble,
-							        std::string assembleMode,
-								bool callFillComplete = true,
-								int FELocExternal=-1);
-
-    void assemblyLaplaceAssFE(int dim,
-                            std::string FEType,
-                            int degree,
-                            int dofs,
-                            BlockMatrixPtr_Type &A,
-                            bool callFillComplete,
-                            int FELocExternal=-1);
-
-    void assemblyAceDeformDiffu(int dim,
-								std::string FETypeChem,
-								std::string FETypeSolid,
-								int degree,
-								int dofsChem,
-								int dofsSolid,
-								MultiVectorPtr_Type c_rep,
-								MultiVectorPtr_Type d_rep,
-								BlockMatrixPtr_Type &A,
-								BlockMultiVectorPtr_Type &resVec,
-								ParameterListPtr_Type params,
-							        std::string assembleMode,
-								bool callFillComplete = true,
-								int FELocExternal=-1);
-
-    void assemblyAceDeformDiffuBlock(int dim,
-                                std::string FETypeChem,
-                                std::string FETypeSolid,
-                                int degree,
-                                int dofsChem,
-                                int dofsSolid,
-                                MultiVectorPtr_Type c_rep,
-                                MultiVectorPtr_Type d_rep,
-                                BlockMatrixPtr_Type &A,
-                                int blockRow,
-                                int blockCol,
-                                BlockMultiVectorPtr_Type &resVec,
-                                int block,
-                                ParameterListPtr_Type params,
-                                std::string assembleMode,
-                                bool callFillComplete = true,
-                                int FELocExternal=-1);
-
-    void advanceInTimeAssemblyFEElements(double dt ,MultiVectorPtr_Type d_rep , MultiVectorPtr_Type c_rep) 
-    {
-        //UN FElocChem = 1; //checkFE(dim,FETypeChem); // Checks for different domains which belongs to a certain fetype
-        UN FElocSolid = 0; //checkFE(dim,FETypeSolid); // Checks for different domains which belongs to a certain fetype
-
-        //ElementsPtr_Type elementsChem= domainVec_.at(FElocChem)->getElementsC();
-
-        ElementsPtr_Type elementsSolid = domainVec_.at(FElocSolid)->getElementsC();
-        
-        vec_dbl_Type solution_c;
-	    vec_dbl_Type solution_d;
-        for (UN T=0; T<assemblyFEElements_.size(); T++) {
-		    vec_dbl_Type solution(0);
-
-            solution_c = getSolution(elementsSolid->getElement(T).getVectorNodeList(), c_rep,1);
-            solution_d = getSolution(elementsSolid->getElement(T).getVectorNodeList(), d_rep,3);
-            // First Solid, then Chemistry
-            solution.insert( solution.end(), solution_d.begin(), solution_d.end() );
-            solution.insert( solution.end(), solution_c.begin(), solution_c.end() );
-            
-            assemblyFEElements_[T]->updateSolution(solution);
-
-            assemblyFEElements_[T]->advanceInTime(dt);
-        }
-        
-    }
-
-	void assemblyLinearElasticity(int dim,
-                                std::string FEType,
-                                int degree,
-                                int dofs,
-                                MultiVectorPtr_Type d_rep,
-                                BlockMatrixPtr_Type &A,
-                                BlockMultiVectorPtr_Type &resVec,
-                                ParameterListPtr_Type params,
-                                bool reAssemble,
-                                std::string assembleMode,
-                                bool callFillComplete=true,
-                                int FELocExternal=-1);
-
-    void assemblyNonLinearElasticity(int dim,
-                                    std::string FEType,
-                                    int degree,
-                                    int dofs,
-                                    MultiVectorPtr_Type d_rep,
-                                    BlockMatrixPtr_Type &A,
-                                    BlockMultiVectorPtr_Type &resVec,
-                                    ParameterListPtr_Type params,
-                                    bool callFillComplete=true,
-                                    int FELocExternal=-1);
-                                    
-    void assemblyNonLinearElasticity(int dim,
-                                    std::string FEType,
-                                    int degree,
-                                    int dofs,
-                                    MultiVectorPtr_Type d_rep,
-                                    BlockMatrixPtr_Type &A,
-                                    BlockMultiVectorPtr_Type &resVec,
-                                    ParameterListPtr_Type params, 									
-                                    DomainConstPtr_Type domain,
-                                    MultiVectorPtr_Type eModVec,
-                                    bool callFillComplete = true,
-                                    int FELocExternal=-1);
-                                    
-    void checkMeshOrientation(int dim, std::string FEType);
-
-    /* Given a converged velocity solution this function 
-       computes the averaged viscosity estimate in each cell at the center of mass 
-       - CM stands for center of mass so the values at the node are averaged to obtain one value
-    */
-    void computeSteadyViscosityFE_CM(int dim,
-	                                    std::string FETypeVelocity,
-	                                    std::string FETypePressure,
-										int dofsVelocity,
-										int dofsPressure,
-										MultiVectorPtr_Type u_rep,
-										MultiVectorPtr_Type p_rep,
- 										ParameterListPtr_Type params);
-
-    // Change for all assembleFEElements the linearization type to the specified ones
-    void  changeLinearizationFE(std::string linearization);
-
-    // Write prostprocessing output fields like e.g. the viscosity based on  velocity, pressure .. solution
-    // inside this BMV -> For visualization or postprocessing                                
-    BlockMultiVectorPtr_Type const_output_fields;
-
-
-
+    Teuchos::RCP<ElementSpec> es_;
 
 /* ----------------------------------------------------------------------------------------*/
 private:
-	void addFeBlockMatrix(BlockMatrixPtr_Type &A, SmallMatrixPtr_Type elementMatrix, FiniteElement element1,FiniteElement element2, MapConstPtr_Type mapFirstColumn,MapConstPtr_Type mapSecondColumn, tuple_disk_vec_ptr_Type problemDisk);
-
-	void addFeBlock(BlockMatrixPtr_Type &A, SmallMatrixPtr_Type elementMatrix, FiniteElement element, MapConstPtr_Type mapFirstRow, int row, int column, tuple_disk_vec_ptr_Type problemDisk);
-
-    void initAssembleFEElements(std::string elementType, tuple_disk_vec_ptr_Type problemDisk, ElementsPtr_Type elements, ParameterListPtr_Type params, vec2D_dbl_ptr_Type pointsRep, MapConstPtr_Type elementMap);
-
-    void addFeBlockMv(BlockMultiVectorPtr_Type &res, vec_dbl_ptr_Type rhsVec, FiniteElement elementBlock1,FiniteElement elementBlock2, int dofs1, int dofs2 );
-
-    void addFeBlockMv(BlockMultiVectorPtr_Type &res, vec_dbl_ptr_Type rhsVec, FiniteElement elementBlock, int dofs);
-		
-    void computeSurfaceNormal(int dim,vec2D_dbl_ptr_Type pointsRep,vec_int_Type nodeList,vec_dbl_Type &v_E, double &norm_v_E);
-
-
-	AssembleFEPtr_vec_Type assemblyFEElements_;
-
-	vec2D_dbl_Type getCoordinates(vec_LO_Type localIDs, vec2D_dbl_ptr_Type points);
-	vec_dbl_Type getSolution(vec_LO_Type localIDs, MultiVectorPtr_Type u_rep, int dofsVelocity);
 
     //Start of AceGen code
     /*! AceGen code for 3D Neo-Hooke material model
@@ -764,77 +572,13 @@ private:
     // void SKR_Biot3D(double* v,ElementSpec *es,ElementData *ed, NodeSpec **ns, NodeData **nd,double *rdata,int *idata,double *p,double **s);
     
     //End of AceGen code
-    
-    void buildTransformation(const vec_int_Type& element,
-                             vec2D_dbl_ptr_Type pointsRep,
-                             SmallMatrix<SC>& B,
-                             std::string FEType="P");
-    
-    void buildTransformation(const vec_int_Type& element,
-                             vec2D_dbl_ptr_Type pointsRep,
-                             SmallMatrix<SC>& B,
-                             vec_dbl_Type& b,
-                             std::string FEType="P");
-
-    
-    void buildTransformationSurface(const vec_int_Type& element,
-                                    vec2D_dbl_ptr_Type pointsRep,
-                                    SmallMatrix<SC>& B,
-                                    vec_dbl_Type& b,
-                                    std::string FEType="P");
 
     void applyDiff(vec3D_dbl_Type& dPhiIn,
                    vec3D_dbl_Type& dPhiOut,
                    SmallMatrix<SC>& diffT);
     
     
-    void phi(       int Dimension,
-                    int intFE,
-            		int i,
-            		vec_dbl_Type &QuadPts,
-            		double* value);
-
-
-    void gradPhi(	int Dimension,
-                    int intFE,
-                    int i,
-                    vec_dbl_Type &QuadPts,
-                    vec_dbl_ptr_Type &value);
-    
-    /*! Most of the quadrature formulas can be found in http://code-aster.org/doc/v11/en/man_r/r3/r3.01.01.pdf 01/2021  */
-    void getQuadratureValues(int Dimension,
-                            int Degree,
-                            vec2D_dbl_ptr_Type &QuadPts,
-                            vec_dbl_ptr_Type &QuadW,
-                            std::string FEType);
-    
-    int getPhi(	vec2D_dbl_ptr_Type &Phi,
-                vec_dbl_ptr_Type &weightsPhi,
-                int Dimension,
-                std::string FEType,
-                int Degree,
-                std::string FETypeQuadPoints="");
-
-    int getPhiGlobal(vec2D_dbl_ptr_Type &Phi,
-                     vec_dbl_ptr_Type &weightsPhi,
-                     int Dimension,
-                     std::string FEType,
-                     int Degree);
-
-    int getDPhi(	vec3D_dbl_ptr_Type &DPhi,
-                	vec_dbl_ptr_Type &weightsDPhi,
-                    int Dimension,
-                    std::string FEType,
-                    int Degree);
-
-    int checkFE(int Dimension,
-                std::string FEType);
-    
-
-    bool setZeros_;
-    SC myeps_;
     std::vector<Teuchos::RCP<DataElement> > ed_;
-    bool saveAssembly_;
 };
 }
 #endif
